# app/Models/Phone.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Factories\HasFactory;
use Illuminate\Database\Eloquent\Model;
use Illuminate\Database\Eloquent\Relations\BelongsTo;
class Phone extends Model
{
    use HasFactory;

    public function organization(): BelongsTo
    {
        return $this->belongsTo(related: Organization::class);
    }
}