# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Organization API",
        "contact": {
            "name": "Данила",
            "url": "https://github.com/DaniiSimo"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        }
    ],
    "paths": {
        "/api/users/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Авторизация",
                "operationId": "4b54274790f7da5a2d3fa7b0d74e6613",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secretpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "$ref": "#/components/schemas/BearerToken"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка аутентификации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": [
                                                "Неправильно указан логин или пароль, повторите попытку."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "email": "Ошибка валидации",
                                                "password": "Ошибка валидации"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Превышен лимит попыток",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Превышен лимит попыток"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": [
                                                "Слишком много попыток входа. Пожалуйста, попробуйте снова через n секунд."
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/{organization}": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Получение организации по идентификатору",
                "operationId": "585449f9b1f0ed989fe5975176b1d18a",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "description": "id организации",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Organization"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка аутентификации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": [
                                                "Доступ запрещён, необходим токен авторизации"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ресурс не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Organization not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Получение списка организаций",
                "operationId": "42efb13746ba71d73c330585f09ecef8",
                "parameters": [
                    {
                        "name": "address_building",
                        "in": "query",
                        "description": "Адрес здания",
                        "schema": {
                            "$ref": "#/components/schemas/Building/properties/address"
                        }
                    },
                    {
                        "name": "name_organization",
                        "in": "query",
                        "description": "Название организации",
                        "schema": {
                            "$ref": "#/components/schemas/Organization/properties/name"
                        }
                    },
                    {
                        "name": "name_activity",
                        "in": "query",
                        "description": "Название деятельности",
                        "schema": {
                            "$ref": "#/components/schemas/Activity/properties/name"
                        }
                    },
                    {
                        "name": "should_take_descendants",
                        "in": "query",
                        "description": "Необходимо ли искать среди потомков деятельности ?(Работает в связке с name_activity)",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "polygon[]",
                        "in": "query",
                        "description": "Координаты области (первая и последняя пара координат должны совпадать)",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "pattern": "^-?\\d{1,3}\\.\\d{2}\\s+-?\\d{1,2}\\.\\d{2}$",
                                "example": "130.15 34.10"
                            },
                            "minItems": 4
                        },
                        "example": [
                            "130.15 34.10",
                            "130.17 34.10",
                            "130.17 34.11",
                            "130.15 34.11",
                            "130.15 34.10"
                        ]
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта, относительно которой будет производиться поиск по радиусу (Работает в связке с lon и radius)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "maximum": 90,
                            "minimum": -90,
                            "example": "34.10"
                        }
                    },
                    {
                        "name": "lon",
                        "in": "query",
                        "description": "Долгота, относительно которой будет производиться поиск по радиусу (Работает в связке с lat и radius)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "double",
                            "maximum": 180,
                            "minimum": -180,
                            "example": "130.15"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус значение в метрах (Работает в связке с lon и lat)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "minimum": 0,
                            "example": "100"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Organization"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка аутентификации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": [
                                                "Доступ запрещён, необходим токен авторизации"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "address_building": "Ошибка валидации",
                                                "name_organization": "Ошибка валидации",
                                                "polygon": "Ошибка валидации",
                                                "lat": "Ошибка валидации",
                                                "lon": "Ошибка валидации",
                                                "radius": "Ошибка валидации"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/registration": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Регистрация",
                "operationId": "bbfb81cce1cc45e09fafe5e937340443",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secretpassword"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Ошибка валидации"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": "Ошибка валидации",
                                                "email": "Ошибка валидации",
                                                "password": "Ошибка валидации"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Activity": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "quo dolor"
                    },
                    "path": {
                        "description": "Путь по которому происходит поиск активности строится через trim(string: Str::slug(title: Str::ascii(value: $value),separator: '_'))",
                        "type": "string",
                        "example": "quo_dolor"
                    },
                    "level": {
                        "description": "Поле выражения, которое описывает глубину деятельности",
                        "type": "smallint",
                        "maximum": 3,
                        "minimum": 1,
                        "readOnly": true,
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "BearerToken": {
                "properties": {
                    "token": {
                        "type": "string",
                        "example": "1|fdsagfdasfdasfdsafdasf"
                    }
                },
                "type": "object"
            },
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "ПАО ГаражХозТелеком"
                    },
                    "lon": {
                        "type": "number",
                        "format": "double",
                        "maximum": 180,
                        "minimum": -180,
                        "example": 130.158708
                    },
                    "lat": {
                        "type": "number",
                        "format": "double",
                        "maximum": 90,
                        "minimum": -90,
                        "example": 34.104894
                    }
                },
                "type": "object"
            },
            "Organization": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "ПАО ГаражХозТелеком"
                    },
                    "building": {
                        "$ref": "#/components/schemas/Building"
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "activites": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Activity"
                        }
                    }
                },
                "type": "object"
            },
            "Phone": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "phone": {
                        "description": "Формат: X-XXX-XXX(-XX-XX). Уникален внутри организации.",
                        "type": "string",
                        "maxLength": 15,
                        "minLength": 9,
                        "pattern": "^[1-9]-[0-9]{3}-[0-9]{3}(-[0-9]{2}-[0-9]{2})?$",
                        "example": "8-923-666-13-13"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "test"
                    },
                    "email": {
                        "type": "email",
                        "example": "test@test.com"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Вводите токен без \"Bearer \" — UI подставит префикс сам",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "User",
            "description": "User"
        },
        {
            "name": "Organization",
            "description": "Organization"
        }
    ]
}